chore: 添加项目工程化配置与代码质量修复
- 新增 .editorconfig、.gitattributes、.pre-commit-config.yaml、cliff.toml
- 新增 .github/workflows/release.yml（GitHub 自动 Release）
- 更新 .gitlab-ci.yml（GitLab 自动 Release）
- 更新 CONTRIBUTING.md（补充 conventional commits 和 git-cliff 说明）
- 全量 ruff 代码质量修复（B904/B007/B905/RUF012/RUF013 等约 190 处）

+++ packages/agentsociety2/agentsociety2/agent/skills/cognition/scripts/update_cognition.py
@@ -18,7 +18,7 @@ def clamp(value: float, lower: float = 0.0, upper: float = 1.0) -> float:
 
 
 def clamp10(value: float) -> int:
-    return int(round(max(0.0, min(10.0, value))))
+    return round(max(0.0, min(10.0, value)))
 
 
 def parse_args() -> argparse.Namespace:
